DEP, MAINT, UPD: update dependencies, update corresponding code
Old code was designed with JAVA 8 and older in mind. That also included usage of libraries and APIs that were outdated/deprecated **already** at the time of development. 
This commit adds changes to pom.xml to reflect all the progress of the current-day JAVA software. Some code had to be changed due to new APIs, but without changing the functionality. 
**NOTE**: since there was/is no tests and corresponding documentation, use this build with caution.